+++ cvxpy/utilities/key_utils.py
@@ -20,6 +20,7 @@ along with CVXPY.  If not, see <http://www.gnu.org/licenses/>.
 # Utility functions to handle indexing/slicing into an expression.
 
 import numpy as np
+import numbers
 
 def validate_key(key, shape):
     """Check if the key is a valid index.
@@ -60,13 +61,26 @@ def format_slice(key_val, dim):
         A slice with a start and step.
     """
     if isinstance(key_val, slice):
+        key_val = slice(to_int(key_val.start),
+                        to_int(key_val.stop),
+                        to_int(key_val.step))
         return key_val
     else:
+        # Convert to int.
+        key_val = to_int(key_val)
         key_val = wrap_neg_index(key_val, dim)
         if 0 <= key_val < dim:
             return slice(key_val, key_val + 1, 1)
         else:
             raise IndexError("Index/slice out of bounds.")
+
+def to_int(val):
+    """Convert everything but None to an int.
+    """
+    if val is None:
+        return val
+    else:
+        return int(val)
 
 def wrap_neg_index(index, dim):
     """Converts a negative index into a positive index.
@@ -143,3 +157,21 @@ def to_str(key):
     """Converts a key (i.e. two slices) into a string.
     """
     return (slice_to_str(key[0]), slice_to_str(key[1]))
+
+def is_special_slice(key):
+    """Does the key contain a list, ndarray, or logical ndarray?
+    """
+    # Key is either a tuple of row, column keys or a single row key.
+    if isinstance(key, tuple):
+        if len(key) > 2:
+            raise IndexError("Invalid index/slice.")
+        key_elems = [key[0], key[1]]
+    else:
+        key_elems = [key]
+
+    # Slices and int-like numbers are fine.
+    for elem in key_elems:
+        if not (isinstance(elem, (numbers.Number, slice)) or np.isscalar(elem)):
+            return True
+
+    return False