+++ examples/extensions/mixed_integer/admm_problem.py
@@ -29,15 +29,17 @@ def admm(self, rho=0.5, iterations=5, *args, **kwargs):
         if getattr(var, "noncvx", False):
             noncvx_vars += [var]
     # Form ADMM problem.
-    obj = self.objective._expr
+    obj = self.objective.args[0]
     for var in noncvx_vars:
         obj = obj + (rho/2)*cvx.sum_entries(cvx.square(var - var.z + var.u))
     prob = cvx.Problem(cvx.Minimize(obj), self.constraints)
     # ADMM loop
     for i in range(iterations):
         result = prob.solve(*args, **kwargs)
-        for var in noncvx_vars:
+        print "relaxation", result
+        for idx, var in enumerate(noncvx_vars):
             var.z.value = var.round(var.value + var.u.value)
+            # print idx, var.z.value, var.value, var.u.value
             var.u.value += var.value - var.z.value
     return polish(self, noncvx_vars, *args, **kwargs)
 
@@ -48,7 +50,7 @@ def admm2(self, rho=0.5, iterations=5, *args, **kwargs):
         if getattr(var, "noncvx", False):
             noncvx_vars += [var]
     # Form ADMM problem.
-    obj = self.objective._expr
+    obj = self.objective.args[0]
     for var in noncvx_vars:
         obj = obj + (rho/2)*cvx.sum_entries(cvx.square(var - var.z + var.u))
     prob = cvx.Problem(cvx.Minimize(obj), self.constraints)